Add abstract phase manager. Turn public processor into a state machine to progress through the phases: 1. Fee Prep 2. App Logic 3. Fee Distribution
Presently the Fee Prep and Fee Distribution phases are noops.
